requirements.txt
Django>=2.2.9,<3
django-bootstrap4-form
celery
django-celery-beat
email-reply-parser
akismet
markdown
beautifulsoup4
lxml
simplejson
pytz
six
djangorestframework
django-model-utils
gevent
gunicorn
whitenoise
#environ
psycopg2
django-anymail
sendgrid
azure-storage
django-storages
django-reversion==3.0.7
django-import-export
django-tinymce

git+https://github.com/darklow/django-suit.git

#pip install https://github.com/darklow/django-suit/tarball/v2